requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.0
groq==0.22.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-google-genai==2.1.3
langchain-neo4j==0.4.0
langchain-openai==0.3.13
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.4.0
langgraph-checkpoint==2.0.25
langgraph-checkpoint-mongodb==0.1.3
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.64
langsmith==0.3.31
marshmallow==3.26.1
mem0ai==0.1.93
monotonic==1.6
motor==3.7.0
multidict==6.4.3
mypy-extensions==1.0.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
numpy==2.2.4
ollama==0.4.7
openai==1.74.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
portalocker==2.10.1
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pymongo==4.12.1
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.13.3
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250402
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
